requirements.txt
Django==4.2.7
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-decouple==3.8
python-dotenv==1.1.1
Pillow==10.1.0
numpy>=1.24.0,<2.0.0
opencv-python-headless==4.8.1.78
openai==1.102.0
openpyxl==3.1.5
python-docx==1.2.0
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
requests==2.31.0

# DeepFace Face Recognition (memory-optimized)
pandas>=2.0.3
tf-keras>=2.20.0

boto3==1.34.30

# Memory optimization - use TensorFlow CPU-only version if available
# tensorflow-cpu  # Uncomment if memory issues persist